Refactor profile validation. Faster. Fixes.
